# ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quantix",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tenders/{tender_id}/boq": {
      "get": {
        "tags": [
          "boq"
        ],
        "summary": "Get Boq",
        "operationId": "get_boq_tenders__tender_id__boq_get",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Boq"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenders/{tender_id}/gates": {
      "get": {
        "tags": [
          "boq"
        ],
        "summary": "Gates",
        "operationId": "gates_tenders__tender_id__gates_get",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gates"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boq/{item_id}/decision": {
      "post": {
        "tags": [
          "boq"
        ],
        "summary": "Decide Item",
        "operationId": "decide_item_boq__item_id__decision_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenders/{tender_id}/boq/approve-all": {
      "post": {
        "tags": [
          "boq"
        ],
        "summary": "Approve All",
        "operationId": "approve_all_tenders__tender_id__boq_approve_all_post",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tender Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approved"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facts/{fact_id}/decision": {
      "post": {
        "tags": [
          "boq"
        ],
        "summary": "Decide Fact",
        "operationId": "decide_fact_facts__fact_id__decision_post",
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fact Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Approved": {
        "properties": {
          "approved": {
            "type": "integer",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "approved"
        ],
        "title": "Approved"
      },
...
      "Boq": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactOut"
            },
            "type": "array",
            "title": "Facts"
          }
        },
        "type": "object",
        "required": [
          "items",
          "facts"
        ],
        "title": "Boq"
      },
...
      "DecisionIn": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "DecisionIn"
      },
...
      "FactOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "label",
          "value",
          "status",
          "proposed_by",
          "reviewed_by",
          "review_note",
          "source"
        ],
        "title": "FactOut"
      },
...
      "Gates": {
        "properties": {
          "manager": {
            "type": "integer",
            "title": "Manager"
          },
          "boq": {
            "type": "integer",
            "title": "Boq"
          },
          "facts": {
            "type": "integer",
            "title": "Facts"
          },
          "takeoff": {
            "type": "integer",
            "title": "Takeoff"
          },
          "drawings": {
            "type": "integer",
            "title": "Drawings"
          },
          "pricing": {
            "type": "integer",
            "title": "Pricing"
          },
          "subcontract": {
            "type": "integer",
            "title": "Subcontract"
          },
          "enquiries": {
            "type": "integer",
            "title": "Enquiries"
          },
          "submission": {
            "type": "integer",
            "title": "Submission"
          }
        },
        "type": "object",
        "required": [
          "manager",
          "boq",
          "facts",
          "takeoff",
          "drawings",
          "pricing",
          "subcontract",
          "enquiries",
          "submission"
        ],
        "title": "Gates"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "item": {
            "type": "string",
            "title": "Item"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "section",
          "item",
          "description",
          "unit",
          "quantity",
          "status",
          "proposed_by",
          "reason",
          "reviewed_by",
          "review_note",
          "source"
        ],
        "title": "ItemOut"
      },
...
      "Source": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "document_name",
          "page",
          "quote"
        ],
        "title": "Source"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}